refactoring on naming
