+++ internal/gpg_key_verifier/gpg_key_verifier.go
@@ -1,4 +1,4 @@
-package gpg_verifier
+package gpg_key_verifier
 
 import (
 	"fmt"
