+++ internal/gpg_key_verifier/gpg_key_verifier_validate.go
@@ -1,8 +1,6 @@
-package gpg_verifier
+package gpg_key_verifier
 
 import (
-	"fmt"
-
 	"github.com/ProtonMail/gopenpgp/v2/crypto"
 )
 
@@ -15,13 +13,4 @@ func (gv gpgKeyVerifier) ValidateSignature(data []byte, signature []byte) error 
 	}
 
 	return nil
-}
-
-func parseKeyRing(pubKeyObj *crypto.Key) (*crypto.KeyRing, error) {
-	signingKeyRing, err := crypto.NewKeyRing(pubKeyObj)
-	if err != nil {
-		return nil, fmt.Errorf("could not build keyring: %w", err)
-	}
-
-	return signingKeyRing, nil
 }
